fix: use integer span counters

+++ packages/loro/src/server/sql.ts
@@ -19,8 +19,8 @@ export const loroSql = `
     document_id UUID NOT NULL,
     update_id UUID NOT NULL,
     peer_id TEXT NOT NULL,
-    start_counter BIGINT NOT NULL,
-    end_counter BIGINT NOT NULL
+    start_counter INTEGER NOT NULL,
+    end_counter INTEGER NOT NULL
   );
   CREATE INDEX IF NOT EXISTS document_update_spans_peer_counter_idx
     ON document_update_spans (document_id, peer_id, end_counter);
